django_site/requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.2
django-bootstrap-form==3.3
django-heroku==0.3.1
djangorestframework==3.7.7
idna==2.6
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.3
requests==2.18.4
urllib3==1.22
whitenoise==3.3.1
